+++ taskmony-front/src/components/edited/edited-task.tsx
@@ -1,6 +1,5 @@
 import yes from "../../images/checkbox-yes.svg";
 import no from "../../images/checkbox-no.svg";
-import follow from "../../images/followed.svg";
 import divider from "../../images/divider.svg";
 import commentsI from "../../images/comment2.svg";
 import createdByI from "../../images/by.svg";
@@ -51,9 +50,9 @@ type TaskProps = {
   createdBy?: string;
   direction?: string;
   save: Function;
-  close: Function;
+  close?: Function;
   changeCheck: Function;
-  deleteTask: Function;
+  deleteTask?: Function;
 };
 
 export const EditedTask = ({
@@ -98,7 +97,7 @@ export const EditedTask = ({
           <img
             src={deleteI}
             onClick={() => {
-              deleteTask(task);
+              if(deleteTask) deleteTask(task);
             }}
             className={"deleteBtn w-4 mt-1 mr-1 cursor-pointer"}
           />
@@ -106,7 +105,7 @@ export const EditedTask = ({
           <img
             src={closeI}
             onClick={() => {
-              close();
+              if(close) close();
               dispatch({ type: RESET_TASK });
             }}
             className={"w-4 p-0.5 cursor-pointer"}
@@ -114,8 +113,8 @@ export const EditedTask = ({
         )}
       </div>
       <Description />
-      <Details recurrent={recurrent} fromDirection={direction} />
-      {task.id && <Comments comments={task.comments} taskId={task.id} />}
+      <Details fromDirection={direction} />
+      {task.id && <Comments />}
       <div className={"w-full flex justify-end"}>
         {!task.id ? (
           <BigBtn
@@ -157,7 +156,7 @@ const Description = () => {
             dispatch(
               changeTaskDetails(
                 taskId,
-                e.target.value == "" ? null : e.target.value
+                e.target.value
               )
             );
         }}
@@ -187,8 +186,8 @@ const Description = () => {
     </div>
   );
 };
-
-const Details = ({ recurrent, fromDirection }) => {
+type DetailsProps = {fromDirection?: string | null}
+const Details = ({ fromDirection } : DetailsProps) => {
   const dispatch = useAppDispatch();
   const task = useAppSelector((store) => store.editedTask);
   const repeatOptions = ["no", "daily", "custom"];
@@ -196,7 +195,7 @@ const Details = ({ recurrent, fromDirection }) => {
     (i) => i.deletedAt == null
   );
 
-  const repeatModeTranslator = (mode) => {
+  const repeatModeTranslator = (mode : string) => {
     switch (mode) {
       case "no":
         return null;
@@ -210,16 +209,14 @@ const Details = ({ recurrent, fromDirection }) => {
         return "daily";
       case "WEEK":
         return "custom";
+      default: return null;
     }
+    return null;
   };
   const defaultUntilDate = () => {
     const date = new Date(Date.now());
-    return date.setFullYear(date.getFullYear() + 1);
+    return new Date(date.setFullYear(date.getFullYear() + 1));
   };
-
-  const customPicker = (
-    <NumberPicker title={"every"} min={2} max={10} after={"days"} hasBorder />
-  );
   const members = directions.filter((d) => d.id == task.direction?.id)[0]
     ?.members;
 
@@ -230,11 +227,11 @@ const Details = ({ recurrent, fromDirection }) => {
           title={"direction"}
           option={task.direction?.name ? task.direction?.name : "none"}
           options={["none", ...directions.map((dir) => dir.name)]}
-          onChange={(index) => {
+          onChange={(index : number) => {
             console.log(index);
             const payload = index == 0 ? null : directions[index - 1];
             dispatch({ type: CHANGE_TASK_DIRECTION, payload: payload });
-            if (task.id) dispatch(changeTaskDirection(task.id, payload));
+            if (task.id && payload) dispatch(changeTaskDirection(task.id, payload));
           }}
           hasBorder
         />
@@ -244,7 +241,7 @@ const Details = ({ recurrent, fromDirection }) => {
           title={"assignee"}
           options={members.map((m) => m.displayName)}
           option={task.assignee?.displayName ? task.assignee?.displayName : "undefined"}
-          onChange={(index) => {
+          onChange={(index : number) => {
             console.log(index);
             const payload = members[index];
             console.log(payload);
@@ -259,7 +256,7 @@ const Details = ({ recurrent, fromDirection }) => {
         title={"start date"}
         date={task.startAt ? new Date(task.startAt) : new Date()}
         hasBorder
-        onChange={(value) => {
+        onChange={(value  : string) => {
           dispatch({ type: CHANGE_TASK_START_DATE, payload: value });
           if (task.id) dispatch(changeTaskStartAt(task.id, value));
         }}
@@ -267,8 +264,8 @@ const Details = ({ recurrent, fromDirection }) => {
       <ItemPicker
         title={"repeated"}
         options={repeatOptions}
-        option={repeatModeTranslator(task.repeatMode)}
-        onChange={(index) => {
+        option={repeatModeTranslator(task.repeatMode) || "no"}
+        onChange={(index : number) => {
           dispatch({
             type: CHANGE_TASK_REPEAT_MODE,
             payload: repeatModeTranslator(repeatOptions[index]),
@@ -312,7 +309,7 @@ const Details = ({ recurrent, fromDirection }) => {
             task.repeatUntil ? new Date(task.repeatUntil) : defaultUntilDate()
           }
           hasBorder
-          onChange={(value) => {
+          onChange={(value : string) => {
             dispatch({ type: CHANGE_TASK_REPEAT_UNTIL, payload: value });
             if (task.id) dispatch(changeTaskRepeatUntil(task.id, value));
           }}
@@ -326,7 +323,7 @@ const Details = ({ recurrent, fromDirection }) => {
             max={9}
             after={"week(s)"}
             value={task.repeatEvery}
-            onChange={(value) => {
+            onChange={(value  : string) => {
               console.log(value);
               dispatch({ type: CHANGE_TASK_REPEAT_EVERY, payload: value });
               if (task.id)
@@ -343,7 +340,7 @@ const Details = ({ recurrent, fromDirection }) => {
           />
           <WeekPicker
             value={task.weekDays}
-            onChange={(value) => {
+            onChange={(value : Array<string>) => {
               dispatch({ type: CHANGE_TASK_REPEAT_WEEK_DAYS, payload: value });
               if (task.id)
                 dispatch(
